// clinic-angular/src/app/models/user.ts
import { Role } from './role';

export class User {
  // userId?: number;
  // firstName?: string;
  // lastName?: string;
  // contactNo?: string;
  // email?: string;
  // password?: string;
  // address?: string;
  // roles?: any;
  // enabled?: string;
  // activationToken?: string;

  userId?: number;
  firstName?: string;
  lastName?: string;
  contactNo?: string;
  email?: string;
  address?: string;
  roles?: any;
  enabled?: string;
  activationToken?: string;
}